pyproject.toml
[project]
name = "lattice-jepa"
version = "0.1.0"
description = "Self-Supervised Pretraining of Recurrent Neural Quantum States for Strongly Correlated Electron Systems"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "pytest>=9.0.2",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" },
]
